Fix clock sync when bicycle started recording first
If the bicycle log starts recording first, the sync flag will occur
sooner (relative to the start of log) in the LIDAR log. As the LIDAR log
is always assumed to be longer, time synchronization using 'valid' mode
with np.correlate() will fail without padding the front of the time and
sync signals.