Fix classification of stationary clusters
Occasionally, a cluster will contain stationary points in addition to
a portion of the cyclist trajectory. These can be identified when the
zmean of the cluster is not near the zmidpoint of the event.

This commit split these clusters. It looks at the xy bounding box for
the cluster in the first and second half of the event and retains points
contained within the smaller bounding box. Excluded points (for all
split clusters if there are more than one) are placed in a new
stationary cluster.